Refactor date and hex color functions to use util.js
