+++ lib/data-generator.js
@@ -2,6 +2,8 @@ var Chance = require('chance');
 var _ = require('lodash');
 var traverse = require('traverse');
 
+var Util = require('./util');
+
 function DataGenerator(context) {
     this.context = context;
     this.chance = new Chance(context.seed);
@@ -27,89 +29,23 @@ DataGenerator.prototype.stylesheet = function () {
 DataGenerator.prototype.date = function (format) {
     var date;
 
-    function getfullDay(day){
-        var fullDay;
-        switch(day) {
-            case 'Mon':
-                fullDay = 'Monday, ';
-                break;
-            case 'Tue':
-                fullDay = 'Tuesday, ';
-                break;
-            case 'Wed':
-                fullDay = 'Wednesday, ';
-                break;
-            case 'Thu':
-                fullDay = 'Thursday, ';
-                break;
-            case 'Fri':
-                fullDay = 'Friday, ';
-                break;
-            case 'Sat':
-                fullDay = 'Saturday, ';
-                break;
-            default:
-                fullDay = 'Sunday, ';
-        }
-
-        return fullDay;
-    }
-
-    function getfullMonth(month){
-        var fullMonth;
-        switch(month){
-            case 'Jan':
-                fullMonth = "January ";
-                break;
-            case 'Feb':
-                fullMonth = "February ";
-                break;
-            case 'Mar':
-                fullMonth = "March ";
-                break;
-            case 'Apr':
-                fullMonth = "April ";
-                break;
-            case 'May':
-                fullMonth = "May ";
-                break;
-            case 'Jun':
-                fullMonth = "June ";
-                break;
-            case 'Jul':
-                fullMonth = "July ";
-                break;
-            case 'Aug':
-                fullMonth = "August ";
-                break;
-            case 'Sep':
-                fullMonth = "September ";
-                break;
-            case 'Oct':
-                fullMonth = "October ";
-                break;
-            case 'Nov':
-                fullMonth = "November ";
-                break;
-            default:
-                fullMonth = "December ";
-                break;
-            }
-        return fullMonth;
-    }
-
-    if(format == 'long') {
+    if(format == 'unformatted') {
+        date = this.chance.date();
+    } else if(format == 'short') {
+        date = this.chance.date({string: true});
+    } else if(format == 'iso') {
+        var iso = this.chance.date({string: true});
+            iso = iso.split("/");
+            date = iso[2] + '-' + iso[0] + '-' + iso[1];
+    } else {
         var dateString, dateArray;
         dateString = this.chance.date();
         dateString = dateString.toString();
         dateString = dateString.substring(0,15);
         dateString = dateString.split(" ");
-        for(i=0; i<4; i++){
-            if(i == 0){
-                dateArray = getfullDay(dateString[i]);
-            } else if(i == 1) {
-                dateArray += getfullMonth(dateString[i]);
-
+        for(i=1; i<4; i++){
+            if(i == 1) {
+                dateArray = Util.getfullMonth(dateString[i]);
             } else if(i == 2) {
                 dateArray += dateString[i] + ', ';
             }
@@ -118,99 +54,21 @@ DataGenerator.prototype.date = function (format) {
             }
         }
         date = dateArray;
-    } else if(format == 'short') {
-        date = this.chance.date({string: true});
-    } else if(format == 'iso') {
-        var iso = this.chance.date({string: true});
-            iso = iso.split("/");
-            date = iso[2] + '-' + iso[0] + '-' + iso[1];
-    } else if(!format || format == 'full') {
-        date = this.chance.date();
     }
     return date;
 };
 
 DataGenerator.prototype.hexColor = function(luminosity) {
-
     var color, hex, rgb, hsl;
-
-    function hex2rgb(hex){
-        hex = hex.replace('#','');
-        var r = parseInt(hex.substring(0,2), 16),
-            g = parseInt(hex.substring(2,4), 16),
-            b = parseInt(hex.substring(4,6), 16);
-
-        return [r, g, b];
-    }
-
-    function rgb2hsl(r, g, b){
-        r /= 255, g /= 255, b /= 255;
-        var max = Math.max(r, g, b), min = Math.min(r, g, b);
-        var h, s, l = (max + min) / 2;
-
-        if(max == min){
-            h = s = 0; // achromatic
-        } else {
-            var d = max - min;
-            s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-            switch(max){
-                case r: h = (g - b) / d + (g < b ? 6 : 0); break;
-                case g: h = (b - r) / d + 2; break;
-                case b: h = (r - g) / d + 4; break;
-            }
-            h /= 6;
-        }
-
-        return [h, s, l];
-    }
-
-    function hsl2rgb(h, s, l){
-        var r, g, b;
-
-        if(s == 0){
-            r = g = b = l; // achromatic
-        } else {
-            function hue2rgb(p, q, t){
-                if(t < 0) t += 1;
-                if(t > 1) t -= 1;
-                if(t < 1/6) return p + (q - p) * 6 * t;
-                if(t < 1/2) return q;
-                if(t < 2/3) return p + (q - p) * (2/3 - t) * 6;
-                return p;
-            }
-
-            var q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-            var p = 2 * l - q;
-            r = hue2rgb(p, q, h + 1/3);
-            g = hue2rgb(p, q, h);
-            b = hue2rgb(p, q, h - 1/3);
-        }
-
-        return [r * 255, g * 255, b * 255];
-    }
-
-
-    function rgb2hex(rgba){
-      var values = rgba
-        .replace(/rgba?\(/, '')
-        .replace(/\)/, '')
-        .replace(/[\s+]/g, '')
-        .split(',');
-      var a = parseFloat(values[3] || 1),
-          r = Math.floor(a * parseInt(values[0]) + (1 - a) * 255),
-          g = Math.floor(a * parseInt(values[1]) + (1 - a) * 255),
-          b = Math.floor(a * parseInt(values[2]) + (1 - a) * 255);
-      return "#" + ("0" + r.toString(16)).slice(-2) + ("0" + g.toString(16)).slice(-2) + ("0" + b.toString(16)).slice(-2);
-    }
 
     luminosity = this.number(luminosity);
 
     if(luminosity === 0 || luminosity) {
         hex = this.chance.color({format: 'hex'});
-        rgb = hex2rgb(hex);
-        hsl = rgb2hsl(rgb[0], rgb[1], rgb[2]);
-        rgb = hsl2rgb(hsl[0], hsl[1], luminosity/100);
-        color = rgb2hex('rgba('+rgb[0]+', '+ rgb[1]+', '+ rgb[2]+')');
+        rgb = Util.hex2rgb(hex);
+        hsl = Util.rgb2hsl(rgb[0], rgb[1], rgb[2]);
+        rgb = Util.hsl2rgb(hsl[0], hsl[1], luminosity/100);
+        color = Util.rgb2hex('rgba('+rgb[0]+', '+ rgb[1]+', '+ rgb[2]+')');
     } else {
         color = this.chance.color({format: 'hex'});
     }
